Add internal transition feature
An internal transition is similar to a normal transition, except it
does not leave the state in which it is defined. Thus the entry, do
and exit actions are not executed when the transition is triggered.
The ability to assign the trigger event, action and guard remain
the same as a normal transition.