Simply scope implementation by leveraging state context hierarchy
When each state is initialized, it requires knowledge of it's parent
context. We can use this knowledge to extend the parent context's
scope with a new child scope using ChainMap.
Now that there is a hierarchy of ChainMap's we need to add a deep
ChainMap implementation - called Scope in order to write to attributes
within nested scopes. Read access is provided by the default ChainMap
implementation.

This also solves a problem with the previous implementation in that the
hierarchy of the scopes match that of the statechart. As a result,
scopes flow down, therefore parent scopes cannot access attributes of
child scopes.