Implement Choice PseudoState
The Choice pseudo-state is used to compose complex transitional path which,
which, when reached, result in the dynamic evaluation of the guards of the
triggers of its outgoing transitions.

Add scope variable to actions & guards so they can read state variables from
the currently active state(s).

Add Else Guard class, useful for associating with outgoing transitions from
choice states to ensure there is always a valid outgoing transition.