Fix default transition from final composite state.
Once a transition is dispatched to a final state, the state will
attempt to dispatch outgoing default transitions. When dispatching
default transitions, the event trigger of the transition to the
final state needs to be cleared.

Allow events to be dispatched to the final state.
This scenario is valid, as the final state is not a transient
pseudostate. The statechart may remain in a final state if a default
transition is blocked or if the composite state is a region within a
compound state. As a transition cannot be explicitly added to a final
state, the parent state will become responsible for dispatching the
transition.